requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI and Language Models
openai>=1.3.7
langchain>=0.0.348
langchain-openai>=0.0.2
langchain-community>=0.0.10

# Document processing and chunking (AI-optimized)
docling==2.52.0          # Advanced document processing for AI workflows
tiktoken==0.11.0         # OpenAI tokenizer for chunking
pypdf==3.17.4           # PDF backup processing
python-docx==1.1.0      # Word document processing
beautifulsoup4==4.12.2  # HTML processing
lxml==5.4.0             # XML/HTML parsing

# Database and Vector Search
psycopg[binary]==3.2.10  # PostgreSQL adapter (Python 3.13 compatible)
sqlalchemy==2.0.23       # Database ORM
chromadb==0.4.22         # Vector database (embedded)
numpy<2.0                # ChromaDB compatibility

# HTTP requests and API communication
httpx==0.25.2
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Data processing (compatible with Python 3.13)
pandas>=2.2.0
numpy>=1.26.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS for frontend integration
python-multipart==0.0.6

# JSON processing (using compatible version)
pydantic>=2.0.0,<2.5.0

# Logging
loguru==0.7.2
